Refactor: Extract shared ItemSubmenuIcon component
- Move identical ItemSubmenuIcon from navigation-link/icons.js to navigation-link/shared/icons.js
- Update navigation-submenu/edit.js to import from shared location
- Update page-list-item/edit.js to import from shared location
- Remove duplicate icons.js file from navigation-submenu block
- Remove unused import from navigation-link/edit.js

This is the first step in a larger refactoring effort to reduce duplication between navigation blocks.
